rename repo to `kms-go` and switch to monorepo
This commit renames the repo to `kms-go`, as more general,
and move KES to its own sub-module.

From now on, the `minio/kms-go` repo contains multiple,
at least two, Go modules:
 - `github.com/minio/kms-go/kms` (KMS SDK)
 - `github.com/minio/kms-go/kes` (KES SDK)

Both, KMS and KES SDKs, are closely related but follow
different lifecycles. If we release one, we may not want
to make a release of the other.

Signed-off-by: Andreas Auernhammer <github@aead.dev>