add support for adding/removing secret key versions
A KMS server supports multiple key versions for a single key.
This commit adds client APIs for adding and removing secret
key versions.

Adding a secret key version behaves like an upsert by creating
the key if it does not exist and otherwise adding another key
version to it.
Similarly, removing a secret key version removes the particular
key version (while `0` refers to the latest version) and, if the
key contains just a single version, also deletes the key.

Signed-off-by: Andreas Auernhammer <github@aead.dev>