kms: remove `Enclave` field from object requests
This commit removes the `Enclave` field from various
object request types, like `CreateKeyRequest`.

Instead, the enclave should be specified as part of the
API call directly. For example:
```
client.CreateKey(ctx, "my-enclave", &kms.CreateKeyRequest{})
```

This change is required for potential pipelining support.
All operations within a pipeline have to operate within the
same enclave (or outside of an enclave). A pipeline with multiple
commands that operate on separate enclaves will not be supported.

Hence, the enclave should be specified once outside of the
object request.