add support for tagged identities
This commit adds support for tagged identities.
An identity can have zero, one or multiple tags
(key-value metadata labels). Tags are useful to
associate additional metadata with an identity.
For example, which service is using this identity.

A tag key is an arbitrary string (at most 64 characters long)
and can contain the following characters: [a-z], [A-Z], [0-9] or
'.', '_', '-', '/'.

A tag value is an arbitrary string from the same character set
but it can be long 256 characters long.

An identity can hold up to 32 tags.

Signed-off-by: Andreas Auernhammer <github@aead.dev>