kms: add `DeleteClusterOnHost` option when removing nodes (#27)
This commit adds the `DeleteClusterOnHost` option that can
be specified when removing a node. When set, the node that
gets removed not just removes itself from the cluster definition
but also deletes all cluster information.

This allows removed nodes to restart since they no longer contain
a cluster definition describing a cluster they are not a part of.
Depending on the situation it may or may not desirable to allow
nodes to be restarted. On example may be when removed nodes should
be joined to the same or a different cluster again (auto-scaling).

Signed-off-by: Andreas Auernhammer <github@aead.dev>